+++ src/main/java/me/Tamaninja/test/repository/LocationRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocationRepo extends JpaRepository<Location, Integer> {
+public interface LocationRepo extends JpaRepository<Location, Long> {
 
 }
